Complete implementation of UI Tests
