+++ .gitignore
@@ -18,3 +18,4 @@ DerivedData
 Pods/
 Carthage/
 /build
+/Demo/UITests/
