+++ .sling.yml
@@ -9,5 +9,5 @@ cocoapods:
   tests:
   runs:
     - scheme: Tempura
-      sdk: iphonesimulator11.2
-      simulator: platform=iOS Simulator,OS=11.2,name=iPhone 7
+      sdk: iphonesimulator11.3
+      simulator: platform=iOS Simulator,OS=11.3,name=iPhone 7
