+++ Demo/Info.plist
@@ -40,8 +40,8 @@
 		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
 	<key>UI_TEST_DIR</key>
-	<string>$(SOURCE_ROOT)/Demo/UI-Tests</string>
+	<string>$(SOURCE_ROOT)/Demo/UITests</string>
 	<key>UI_TEST_RECORDING</key>
-	<true/>
+	<false/>
 </dict>
 </plist>
