+++ Tempura.xcodeproj/xcshareddata/xcschemes/Demo.xcscheme
@@ -14,7 +14,7 @@
             buildForArchiving = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "0F48833F1EA5BD00B9B2704D"
+               BlueprintIdentifier = "972F6A2D7EDC209C37F49DD7"
                BlueprintName = "Demo"
                ReferencedContainer = "container:Tempura.xcodeproj"
                BuildableName = "Demo.app">
@@ -28,7 +28,7 @@
             buildForArchiving = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "374A9C0734B7AA0E90ED59A9"
+               BlueprintIdentifier = "75476E32CE4BCBFD2A99320B"
                BlueprintName = "DemoTests"
                ReferencedContainer = "container:Tempura.xcodeproj"
                BuildableName = "DemoTests.xctest">
@@ -48,7 +48,7 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "374A9C0734B7AA0E90ED59A9"
+               BlueprintIdentifier = "75476E32CE4BCBFD2A99320B"
                BlueprintName = "DemoTests"
                ReferencedContainer = "container:Tempura.xcodeproj"
                BuildableName = "DemoTests.xctest">
@@ -72,7 +72,7 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "0F48833F1EA5BD00B9B2704D"
+            BlueprintIdentifier = "972F6A2D7EDC209C37F49DD7"
             BlueprintName = "Demo"
             ReferencedContainer = "container:Tempura.xcodeproj"
             BuildableName = "Demo.app">
@@ -88,7 +88,7 @@
       <BuildableProductRunnable>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "0F48833F1EA5BD00B9B2704D"
+            BlueprintIdentifier = "972F6A2D7EDC209C37F49DD7"
             BlueprintName = "Demo"
             ReferencedContainer = "container:Tempura.xcodeproj"
             BuildableName = "Demo.app">
